requirements.txt
Django==1.6.5
Markdown==2.5.1
PyYAML==3.11
Scrapy==0.24.4
South==1.0
Twisted==14.0.0
argparse==1.2.1
awscli==1.4.2
bcdoc==0.12.2
botocore==0.62.0
-e git://github.com/bufferapp/buffer-python.git@53995bbd067e9fbdbc519ac4a4809e97c82272aa#egg=buffer_python-master
cffi==0.8.6
colorama==0.2.7
coverage==3.7.1
cryptography==0.5.4
cssselect==0.9.1
dj-database-url==0.2.2
dj-static==0.0.5
django-cors-headers==0.13
django-filter==0.8
django-grappelli==2.5.3
django-selectable==0.8.0
djangorestframework==2.4.3
docutils==0.12
gunicorn==18.0
jmespath==0.4.1
lxml==3.3.5
mercurial==3.1.2
mock==1.0.1
nose==1.3.0
pbr==0.10.0
psycopg2==2.5.3
pyOpenSSL==0.14
pyasn1==0.1.7
pycparser==2.10
python-dateutil==2.2
queuelib==1.2.2
rauth==0.6.2
requests==1.2.3
rsa==3.1.2
six==1.8.0
static==0.4
stevedore==0.15
vboxapi==1.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
w3lib==1.10.0
wsgiref==0.1.2
youtube-dl==2014.08.29
zope.interface==4.1.1
